+++ src/main/java/com/dnd12th_4/pickitalki/controller/login/dto/response/RefreshTokenResponse.java
@@ -1,0 +1,16 @@
+package com.dnd12th_4.pickitalki.controller.login.dto.response;
+
+
+import lombok.*;
+
+import java.time.LocalDateTime;
+
+@Data
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+public class RefreshTokenResponse {
+
+    private String accessToken;
+    private LocalDateTime expiredAccessToken;
+}
