+++ src/main/java/com/dnd12th_4/pickitalki/controller/login/dto/response/UserResponse.java
@@ -1,4 +1,4 @@
-package com.dnd12th_4.pickitalki.controller.login.dto;
+package com.dnd12th_4.pickitalki.controller.login.dto.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
