+++ src/main/java/com/dnd12th_4/pickitalki/service/login/KakaoUserService.java
@@ -1,6 +1,6 @@
 package com.dnd12th_4.pickitalki.service.login;
 
-import com.dnd12th_4.pickitalki.controller.login.dto.KakaoUserDto;
+import com.dnd12th_4.pickitalki.controller.login.dto.response.KakaoUserDto;
 import com.dnd12th_4.pickitalki.presentation.error.TokenErrorCode;
 import com.dnd12th_4.pickitalki.presentation.exception.ApiException;
 import com.dnd12th_4.pickitalki.service.login.tool.HttpCreator;
